+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.6.0 - TBD
+- Added support for imbalance schema
 - Changed `Historical::BatchDownload` to return the paths of the downloaded files
 
 ## 0.5.0 - 2023-03-13